refactor: 점수 계산 테스트 위치 변경
