+++ src/main/java/chess/domain/board/ChessBoard.java
@@ -4,7 +4,6 @@ import static chess.domain.board.InitialPieces.INITIAL_PIECES;
 import static chess.domain.piece.Team.WHITE;
 
 import chess.domain.Position;
-import chess.domain.game.PointCalculator;
 import chess.domain.piece.Piece;
 import chess.domain.piece.PieceMoveResult;
 import chess.domain.piece.PieceType;
