+++ src/test/java/chess/domain/board/ChessBoardTest.java
@@ -1,10 +1,8 @@
 package chess.domain.board;
 
 import static chess.domain.Position.A1;
-import static chess.domain.Position.A2;
 import static chess.domain.Position.A3;
 import static chess.domain.Position.A4;
-import static chess.domain.Position.A5;
 import static chess.domain.Position.A6;
 import static chess.domain.Position.A8;
 import static chess.domain.Position.B1;
@@ -30,7 +28,6 @@ import static chess.domain.piece.PieceMoveResult.FAILURE;
 import static chess.domain.piece.PieceMoveResult.SUCCESS;
 import static chess.domain.piece.Team.BLACK;
 import static chess.domain.piece.Team.WHITE;
-import static org.junit.jupiter.api.Assertions.assertAll;
 
 import chess.domain.Position;
 import chess.domain.piece.Bishop;
@@ -46,7 +43,6 @@ import java.util.List;
 import java.util.Optional;
 import java.util.stream.Stream;
 import org.assertj.core.api.Assertions;
-import org.assertj.core.data.Offset;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -169,33 +165,5 @@ class ChessBoardTest {
         Optional<Team> actual = chessBoard.whichTeam(D2);
         Assertions.assertThat(actual)
                 .isEmpty();
-    }
-
-    @Test
-    @DisplayName("각 팀의 점수가 잘 계산되는지 검증")
-    void calculatePoint() {
-        ChessBoard chessBoard = new ChessBoard();
-        double whiteTeamPoint = chessBoard.calculatePoint(WHITE);
-        double blackTeamPoint = chessBoard.calculatePoint(BLACK);
-        Offset<Double> offset = Offset.offset(0.01);
-        assertAll(
-                () -> Assertions.assertThat(whiteTeamPoint).isCloseTo(38.0, offset),
-                () -> Assertions.assertThat(blackTeamPoint).isCloseTo(38.0, offset)
-        );
-    }
-
-    @Test
-    @DisplayName("같은 열의 폰이 포함된 각 팀의 점수가 잘 계산되는지 검증")
-    void calculatePointWithPawn() {
-        ChessBoard chessBoard = new ChessBoard(new King(A8, WHITE), new King(H1, BLACK), new Pawn(A2, BLACK),
-                new Pawn(B2, BLACK), new Pawn(A3, BLACK),
-                new Pawn(A4, WHITE), new Pawn(A5, WHITE), new Pawn(A6, WHITE));
-        double whiteTeamPoint = chessBoard.calculatePoint(WHITE);
-        double blackTeamPoint = chessBoard.calculatePoint(BLACK);
-        Offset<Double> offset = Offset.offset(0.01);
-        assertAll(
-                () -> Assertions.assertThat(blackTeamPoint).isCloseTo(2.0, offset),
-                () -> Assertions.assertThat(whiteTeamPoint).isCloseTo(1.5, offset)
-        );
     }
 }
